+++ edx_badges/credly/serializers.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+
+class BadgeDataModel(serializers.Serializer):  # pylint: disable=abstract-method
+    id = serializers.CharField()
+    state = serializers.CharField()